refactor: Change packages structure, remove unused dependencies
