+++ src/main/java/com/dev/application/domain/Card.java
@@ -1,4 +1,7 @@
-package com.dev.domain;
+package com.dev.application.domain;
+
+import com.dev.application.domain.enums.Figure;
+import com.dev.application.domain.enums.Suit;
 
 public class Card {
