+++ src/main/java/com/dev/application/domain/Person.java
@@ -1,4 +1,4 @@
-package com.dev.domain;
+package com.dev.application.domain;
 
 import java.util.List;
