+++ src/test/java/com/dev/DealerTest.java
@@ -1,6 +1,8 @@
 package com.dev;
 
-import com.dev.domain.*;
+import com.dev.application.domain.*;
+import com.dev.application.domain.enums.Figure;
+import com.dev.application.domain.enums.Suit;
 import jakarta.inject.Inject;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
