+++ src/test/java/com/dev/DeckTest.java
@@ -1,9 +1,9 @@
 package com.dev;
 
-import com.dev.domain.Card;
-import com.dev.domain.Deck;
-import com.dev.domain.Figure;
-import com.dev.domain.Suit;
+import com.dev.application.domain.Card;
+import com.dev.application.domain.Deck;
+import com.dev.application.domain.enums.Figure;
+import com.dev.application.domain.enums.Suit;
 import io.quarkus.test.junit.QuarkusTest;
 import jakarta.inject.Inject;
 import org.junit.jupiter.api.Assertions;
